Retoques finales al proyecto. Listo para presentacion

+++ gestion/views/registros/views.py
@@ -5,14 +5,16 @@ from django.views.decorators.csrf import csrf_exempt
 from django.utils.decorators import method_decorator
 from django.contrib.auth.mixins import LoginRequiredMixin
 from django.urls import reverse_lazy
+from gestion.mixin import ValidatePermissionRequiredMixin
 from gestion.models import Registro
 from gestion.forms import RegistroForm
 
 #======VISTA DE REGISTROS=======#
-class RegistroListarView(LoginRequiredMixin, TemplateView):
+class RegistroListarView(LoginRequiredMixin, ValidatePermissionRequiredMixin, TemplateView):
     model = Registro
     template_name = 'registros/registros.html'
-    
+    permission_required = 'gestion.view_registro'
+
     @method_decorator(csrf_exempt)
     def dispatch(self, request, *args, **kwargs):
         return super().dispatch(request, *args, **kwargs)
@@ -27,12 +29,16 @@ class RegistroListarView(LoginRequiredMixin, TemplateView):
                 for i in registros:
                     data.append(i.toJSON())
             elif action == 'add':
+                if not request.user.has_perm('gestion.add_registro'):
+                    return JsonResponse({'error': 'No tienes permiso para agregar registros.'}, status=403)
                 form = RegistroForm(request.POST)
                 if form.is_valid():
                     form.save()
                 else: 
                     data['error'] = form.errors.as_json()
             elif action =='edit':
+                if not request.user.has_perm('gestion.change_registro'):
+                    return JsonResponse({'error': 'No tienes permiso para editar registros.'}, status=403)
                 registro = Registro.objects.get(pk=request.POST['id'])
                 form= RegistroForm(request.POST, instance=registro)
                 if form.is_valid():
@@ -41,6 +47,8 @@ class RegistroListarView(LoginRequiredMixin, TemplateView):
                     data['error'] = form.errors.as_json()
                 registro.save()
             elif action =='delete':
+                if not request.user.has_perm('gestion.delete_registro'):
+                    return JsonResponse({'error': 'No tienes permiso para eliminar registros.'}, status=403)
                 registro = Registro.objects.get(pk=request.POST['id'])
                 registro.delete()
             else:
